api/requirements.txt
Flask==2.3.1
python-dotenv==1.0.0
google-ads==19.0.0
cryptography==40.0.2
pymongo==4.3.3
Flask-PyMongo==2.3.0
google-auth==2.16.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.8.0
certifi==2022.12.7
click==8.1.3